requirements.txt
Django==1.6.5
git+https://github.com/bartTC/django-attachments.git
South==0.8.4
mock==1.0.1
django-bootstrap3==4.4.1
django-notification==1.1.1
django-ajax-selects==1.3.4
django-bootstrap-pagination==1.5.0
